protect all routes from URL hacking

+++ app/controllers/items_controller.rb
@@ -1,65 +1,93 @@
 class ItemsController < ApplicationController
 
 	get "/items" do
-		@items = Item.order(name: :asc)
-		
-		erb :"/items/index"
+		if logged_in?
+			@items = Item.order(name: :asc)
+			
+			erb :"/items/index"
+		else
+			redirect "/login"
+		end
 	end
 
 	get "/items/new" do
-		@locations = Location.all
+		if logged_in?
+			@locations = Location.all
 
-		erb :"/items/new"
+			erb :"/items/new"
+		else
+			redirect "/login"
+		end
 	end
 
 	post "/items" do
-		item = Item.new(params[:item])
+		if logged_in?
+			item = Item.new(params[:item])
 
-		if item.save
-			flash[:message] = "#{item.name} created!"
+			if item.save
+				flash[:message] = "#{item.name} created!"
 
-			redirect "/items"
+				redirect "/items"
+			else
+				flash[:message] = item.errors.messages.collect {|k,v| "#{k.to_s.capitalize} #{v.join}"}
+
+				redirect "/items/new"
+			end
 		else
-			flash[:message] = item.errors.messages.collect {|k,v| "#{k.to_s.capitalize} #{v.join}"}
-
-			redirect "/items/new"
+			redirect "/login"
 		end
 	end
 
 	get "/items/shopping-list" do
-		@items = Item.order(name: :asc)
+		if logged_in?
+			@items = Item.order(name: :asc)
 
-		erb :"/items/shopping-list"
+			erb :"/items/shopping-list"
+		else
+			redirect "/login"
+		end
 	end
 
 	get "/items/:slug/edit" do
-		@item = Item.find_by_slug(params[:slug])
-		@locations = Location.all
+		if logged_in?
+			@item = Item.find_by_slug(params[:slug])
+			@locations = Location.all
 
-		erb :"/items/edit"
+			erb :"/items/edit"
+		else
+			redirect "/login"
+		end
 	end
 
 	patch "/items/:slug" do
-		item = Item.find_by_slug(params[:slug])
-		slug = item.slug
+		if logged_in?
+			item = Item.find_by_slug(params[:slug])
+			slug = item.slug
 
-		if item.update(params[:item])
-			flash[:message] = "#{item.name} successfully updated!"
-			redirect "/items"
+			if item.update(params[:item])
+				flash[:message] = "#{item.name} successfully updated!"
+				redirect "/items"
+			else
+				flash[:message] = "Item update error, please try again."
+				redirect "/items/#{slug}/edit"
+			end
 		else
-			flash[:message] = "Item update error, please try again."
-			redirect "/items/#{slug}/edit"
+			redirect "/login"
 		end
 	end
 
 	delete "/items/:slug" do
-		item = Item.find_by_slug(params[:slug])
+		if logged_in?
+			item = Item.find_by_slug(params[:slug])
 
-		item.destroy
+			item.destroy
 
-		flash[:message] = "Item successfully deleted."
+			flash[:message] = "Item successfully deleted."
 
-		redirect "/items"
+			redirect "/items"
+		else
+			redirect "/login"
+		end
 	end
 
  end
